CrewAI/requirements.txt
crewai
flask
redis
psycopg2
python-dotenv
requests
langchain-core
uuid